+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][Keep a Changelog] and this project adheres to [Semantic Versioning][Semantic Versioning].
 
-## [Unreleased]
+## [1.0.4] - 2021-10-29
 
 ### Added
 1. Added support for writing UND sections using `glypy.io.glycoct`. The previous implementation
@@ -128,7 +128,7 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 [Semantic Versioning]: https://semver.org/
 
 <!-- Versions -->
-[Unreleased]: https://github.com/mobiusklein/glypy/compare/v1.0.3...HEAD
+[Unreleased]: https://github.com/mobiusklein/glypy/compare/v1.0.4...HEAD
 [Released]: https://github.com/mobiusklein/glypy/releases
 [0.12.6]: https://github.com/mobiusklein/glypy/releases/v0.12.6
 [0.12.7]: https://github.com/mobiusklein/glypy/releases/v0.12.7
@@ -136,3 +136,4 @@ The format is based on [Keep a Changelog][Keep a Changelog] and this project adh
 [1.0.1]: https://github.com/mobiusklein/glypy/releases/v1.0.1
 [1.0.2]: https://github.com/mobiusklein/glypy/releases/v1.0.2
 [1.0.3]: https://github.com/mobiusklein/glypy/releases/v1.0.3
+[1.0.4]: https://github.com/mobiusklein/glypy/releases/v1.0.4